+++ changelog.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [0.14.0] - 2020-07-26
+### Changed
+* Removed workaround for deployment of Azure Integration Runtimes.
+### Added
+* List of filtering rules (Includes/Excludes) can be provided by file
+* Config file allows defining a property to be added or removed
+
 ## [0.13.0] - 2020-07-08
 ### Fixed
 * All lines were being ignored in config file after commented line (#24)